reorganize files and start adding comments
