+++ run-test.sh
@@ -1,5 +1,5 @@
-g++ src/binary-file.cpp src/buffer/buffer-writter.cpp src/buffer-utils.cpp \
-src/db/hash-file.cpp src/paper-block.cpp src/paper.cpp src/parsing/paper-stream.cpp src/parsing/parser.cpp \
+g++ src/binary-file.cpp src/buffer/buffer.cpp src/buffer/buffer-utils.cpp src/db/hash-file.cpp \
+src/paper-block.cpp src/paper.cpp src/parsing/paper-stream.cpp src/parsing/parser.cpp \
 test/hash-file.test.cpp -o hash.test
 
 ./hash.test
