+++ src/binary-file.h
@@ -9,6 +9,8 @@
 
 /**
  * Esta é uma classe para abstrair a leitura e escrita de blocos de um tamanho X em um arquivo binario qualquer.
+ *
+ * A ideia é não usar `fstream` e afins no código.
  */
 class BinaryFile {
    private:
@@ -26,9 +28,18 @@ class BinaryFile {
 
     void close();
 
+    // pré-aloca espaço no arquivo preenchendo com zeros
     void zero_the_file_out();
 
+    /**
+     * Escreve um buffer em um bloco na posição indicada.
+     * O offset em bytes é calculado a partir do block_index multiplicado pelo tamanho do bloco.
+     * */
     void write_block(const char* block_buffer, int block_index);
 
+    /*
+     Método que lê um bloco em uma posicao x e aloca um buffer para esse bloco.
+     Importante notar que a posição do bloco é calculada a partir do index multiplicando pelo tamanho do bloco. Esse
+     parametro **NÃO** é o offset em bytes */
     unsigned char* read_block(int block_index);
 };
