+++ src/buffer/buffer-utils.cpp
