+++ src/buffer/buffer-utils.h
