+++ src/config.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define BLOCK_SIZE 4096  // 4kb
+#define BUCKETS 400000   // 400k
+#define BLOCKS_PER_BUCKET 2
