--- src/consts.hpp
@@ -1,2 +1,0 @@
-#pragma once
-#define BLOCK_SIZE 4096
