+++ src/paper.h
@@ -1,14 +1,15 @@
 #include <iostream>
 #include <string>
 
-#include "buffer-utils.h"
-
 #pragma once
 
 #define LENGTH_HEADERS 6
 #define FIXED_LENGTH 30
 #define REGISTER_HEADER 2
 
+/**
+ * Classe que representa um artigo.
+ * */
 class Paper {
    public:
     unsigned int id;
